empresas/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==2.1.5
django-configurations==2.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
openapi-codec==1.3.2
pipenv==2018.11.26
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pytest==4.1.0
pytest-cov==2.6.0
python-dotenv==0.10.1
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.4.0
whitenoise==4.1.2
wincertstore==0.2
